mavenize project
metouia L&F seems to be abbandoned so its jar is installed manually

closes #1
