+++ build.xml
@@ -5,16 +5,16 @@
 
     <target name="compile">
         <mkdir dir="${build.dir}"/>
-        <javac srcdir="src/" destdir="${build.dir}" optimize="on" debug="off">
+        <javac srcdir="src/main/java" destdir="${build.dir}" optimize="on" debug="off">
             <include name="**/*.java"/>
         </javac>
 		<copy todir="${build.dir}/" overwrite="yes">
-			<fileset dir="src/">
+         <fileset dir="src/main/java">
 				<include name="*.txt"/>
 			</fileset>
 		 </copy>
 		 <copy todir="${build.dir}/icons" overwrite="yes">
-			<fileset dir="src/icons">
+			<fileset dir="src/main/resources/icons">
 				<include name="*.gif"/>
 			</fileset>
 		 </copy>
