--- nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=cba644ff
-build.xml.script.CRC32=74386eec
-build.xml.stylesheet.CRC32=ba5d3624
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=94bef83f
-nbproject/build-impl.xml.script.CRC32=c8383d25
-nbproject/build-impl.xml.stylesheet.CRC32=0c01fd8e@1.43.1.45
