--- nbproject/private/private.properties
@@ -1,9 +1,0 @@
-application.args=
-compile.on.save=false
-do.depend=false
-do.jar=true
-javac.debug=true
-javadoc.preview=true
-user.properties.file=C:\\Users\\akshath\\.netbeans\\7.0\\build.properties
-default.javac.source=1.3
-default.javac.target=1.3
