+++ src/main/java/net/sf/sockettest/MyTrustManager.java
