+++ src/main/java/net/sf/sockettest/Util.java
