+++ src/main/java/net/sf/sockettest/swing/SocketTestClient.java
@@ -3,19 +3,22 @@ package net.sf.sockettest.swing;
 import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
-import javax.swing.event.*;
 import javax.swing.border.*;
+
+import java.security.SecureRandom;
 
 import java.net.*;
 import java.io.*;
 
-import net.sf.sockettest.*;
+import javax.net.ssl.*;
 
+import net.sf.sockettest.*;
 /**
  *
  * @author Akshathkumar Shetty
  */
-public class SocketTestServer extends JPanel /*JFrame*/ {
+public class SocketTestClient extends JPanel {
+    
     private final String NEW_LINE = "\r\n";
     private ClassLoader cl = getClass().getClassLoader();
     public ImageIcon logo = new ImageIcon(
@@ -33,33 +36,32 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
     private JLabel portLabel = new JLabel("Port");
     private JLabel logoLabel = new JLabel("SocketTest v 3.0", logo,
             JLabel.CENTER);
-    private JTextField ipField = new JTextField("0.0.0.0",20);
+    private JTextField ipField = new JTextField("127.0.0.1",20);
     private JTextField portField = new JTextField("21",10);
     private JButton portButton = new JButton("Port");
-    private JButton connectButton = new JButton("Start Listening");
+    private JButton connectButton = new JButton("Connect");
     
-    private JLabel convLabel = new JLabel("Conversation with Client");
-    private Border connectedBorder = BorderFactory.createTitledBorder(new EtchedBorder(),"Connected Client : < NONE >");
+    private JLabel convLabel = new JLabel("Conversation with host");
+    private Border connectedBorder = BorderFactory.createTitledBorder(new EtchedBorder(),"Connected To < NONE >");
     private JTextArea messagesField = new JTextArea();
     
     private JLabel sendLabel = new JLabel("Message");
     private JTextField sendField = new JTextField();
     
     private JButton sendButton = new JButton("Send");
-    private JButton disconnectButton = new JButton("Disconnect");
     private JButton saveButton = new JButton("Save");
     private JButton clearButton = new JButton("Clear");
     
+    private JCheckBox secureButton = new JCheckBox("Secure");
+    private boolean isSecure = false;
     private GridBagConstraints gbc = new GridBagConstraints();
     
     private Socket socket;
-    private ServerSocket server;
-    private SocketServer socketServer;
     private PrintWriter out;
-    
+    private SocketClient socketClient;
     protected final JFrame parent;
     
-    public SocketTestServer(final JFrame parent) {
+    public SocketTestClient(final JFrame parent) {
         //Container cp = getContentPane();
         this.parent = parent;
         Container cp = this;
@@ -80,7 +82,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         
         gbc.weightx = 1.0; //streach
         gbc.gridx = 1;
-        gbc.gridwidth = 3;
+        gbc.gridwidth = 4;
         gbc.fill = GridBagConstraints.HORIZONTAL;
         ActionListener ipListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -131,12 +133,28 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         gbc.gridx = 3;
         gbc.gridwidth = 1;
         gbc.fill = GridBagConstraints.NONE;
-        connectButton.setMnemonic('S');
-        connectButton.setToolTipText("Start Listening");
+        connectButton.setMnemonic('C');
+        connectButton.setToolTipText("Start Connection");
         connectButton.addActionListener(connectListener);
         toPanel.add(connectButton, gbc);
         
-        toPanel.setBorder(BorderFactory.createTitledBorder(new EtchedBorder(),"Listen On"));
+        
+        gbc.weightx = 0.0;
+        gbc.gridy = 1;
+        gbc.gridx = 4;
+        gbc.gridwidth = 1;
+        gbc.fill = GridBagConstraints.NONE;
+        secureButton.setToolTipText("Set Has Secure");
+        secureButton.addItemListener(
+                new ItemListener() {
+            public void itemStateChanged(ItemEvent e) {
+                isSecure = !isSecure;
+            }
+        });
+        toPanel.add(secureButton, gbc);
+        
+        
+        toPanel.setBorder(BorderFactory.createTitledBorder(new EtchedBorder(),"Connect To"));
         topPanel.setLayout(new BorderLayout(10,0));
         topPanel.add(toPanel);
         logoLabel.setVerticalTextPosition(JLabel.BOTTOM);
@@ -173,7 +191,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         gbc.weightx = 0.0;
         gbc.fill = GridBagConstraints.NONE;
         sendButton.setEnabled(false);
-        sendButton.setToolTipText("Send text to client");
+        sendButton.setToolTipText("Send text to host");
         ActionListener sendListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
                 String msg = sendField.getText();
@@ -181,8 +199,8 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
                     sendMessage(msg);
                 else {
                     int value = JOptionPane.showConfirmDialog(
-                            SocketTestServer.this,  "Send Blank Line ?",
-                            "Send Data To Client",
+                            SocketTestClient.this,  "Send Blank Line ?",
+                            "Send Data To Server",
                             JOptionPane.YES_NO_OPTION);
                     if (value == JOptionPane.YES_OPTION)
                         sendMessage(msg);
@@ -192,16 +210,6 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         sendButton.addActionListener(sendListener);
         sendField.addActionListener(sendListener);
         sendPanel.add(sendButton, gbc);
-        ActionListener disconnectListener = new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                disconnect();
-            }
-        };
-        gbc.gridx = 3;
-        disconnectButton.addActionListener(disconnectListener);
-        disconnectButton.setEnabled(false);
-        sendPanel.add(disconnectButton, gbc);
-        
         sendPanel.setBorder(
                 new CompoundBorder(
                 BorderFactory.createEmptyBorder(0,0,0,3),
@@ -223,7 +231,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         gbc.gridx = 1;
         gbc.gridheight = 1;
         gbc.fill = GridBagConstraints.HORIZONTAL;
-        saveButton.setToolTipText("Save conversation with client to a file");
+        saveButton.setToolTipText("Save conversation with host to a file");
         saveButton.setMnemonic('S');
         ActionListener saveListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -235,13 +243,13 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
                 String fileName="";
                 JFileChooser chooser = new JFileChooser();
                 chooser.setCurrentDirectory(new File("."));
-                int returnVal = chooser.showSaveDialog(SocketTestServer.this);
+                int returnVal = chooser.showSaveDialog(SocketTestClient.this);
                 if(returnVal == JFileChooser.APPROVE_OPTION) {
                     fileName=chooser.getSelectedFile().getAbsolutePath();
                     try {
                         Util.writeFile(fileName,text);
                     } catch (Exception ioe) {
-                        JOptionPane.showMessageDialog(SocketTestServer.this,
+                        JOptionPane.showMessageDialog(SocketTestClient.this,
                                 ""+ioe.getMessage(),
                                 "Error saving to file..",
                                 JOptionPane.ERROR_MESSAGE);
@@ -252,7 +260,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         saveButton.addActionListener(saveListener);
         buttonPanel.add(saveButton, gbc);
         gbc.gridy = 1;
-        clearButton.setToolTipText("Clear conversation with client");
+        clearButton.setToolTipText("Clear conversation with host");
         clearButton.setMnemonic('C');
         ActionListener clearListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -280,28 +288,29 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
     
         /*
         public static void main(String args[]) {
-                SocketTestServer stServer=new SocketTestServer();
-                stServer.setTitle("SocketTest Server");
-                stServer.setSize(500,400);
-                Util.centerWindow(stServer);
-                stServer.setDefaultCloseOperation(EXIT_ON_CLOSE);
-                stServer.setIconImage(stServer.logo.getImage());
-                stServer.setVisible(true);
+                SocketTestClient client=new SocketTestClient();
+                client.setTitle("SocketTest Client");
+                //client.pack();
+                client.setSize(500,400);
+                Util.centerWindow(client);
+                client.setDefaultCloseOperation(EXIT_ON_CLOSE);
+                client.setIconImage(client.logo.getImage());
+                client.setVisible(true);
         }
          */
     
-    /////////////////////
-    //action & helper methods
-    /////////////////////
+    /////////////////
+    //action methods
+    //////////////////
     private void connect() {
-        if(server!=null) {
-            stop();
+        if(socket!=null) {
+            disconnect();
             return;
         }
         String ip=ipField.getText();
         String port=portField.getText();
         if(ip==null || ip.equals("")) {
-            JOptionPane.showMessageDialog(SocketTestServer.this,
+            JOptionPane.showMessageDialog(SocketTestClient.this,
                     "No IP Address. Please enter IP Address",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             ipField.requestFocus();
@@ -309,7 +318,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
             return;
         }
         if(port==null || port.equals("")) {
-            JOptionPane.showMessageDialog(SocketTestServer.this,
+            JOptionPane.showMessageDialog(SocketTestClient.this,
                     "No Port number. Please enter Port number",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             portField.requestFocus();
@@ -318,7 +327,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         }
         setCursor(Cursor.getPredefinedCursor(Cursor.WAIT_CURSOR));
         if(!Util.checkHost(ip)) {
-            JOptionPane.showMessageDialog(SocketTestServer.this,
+            JOptionPane.showMessageDialog(SocketTestClient.this,
                     "Bad IP Address",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             ipField.requestFocus();
@@ -330,7 +339,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
         try	{
             portNo=Integer.parseInt(port);
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(SocketTestServer.this,
+            JOptionPane.showMessageDialog(SocketTestClient.this,
                     "Bad Port number. Please enter Port number",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             portField.requestFocus();
@@ -339,86 +348,72 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
             return;
         }
         try {
-            InetAddress bindAddr=null;
-            if(!ip.equals("0.0.0.0"))
-                bindAddr = InetAddress.getByName(ip);
-            else
-                bindAddr = null;
-            server = new ServerSocket(portNo,1,bindAddr);
+            if(isSecure==false) {
+                System.out.println("Connectig in normal mode : "+ip+":"+portNo);
+                socket = new Socket(ip,portNo);
+            } else {
+                System.out.println("Connectig in secure mode : "+ip+":"+portNo);
+                //SocketFactory factory = SSLSocketFactory.getDefault();
+				
+				TrustManager[] tm = new TrustManager[] { new MyTrustManager(SocketTestClient.this) }; 
+
+                SSLContext context = SSLContext.getInstance("TLS");
+                context.init(new KeyManager[0], tm, new SecureRandom());
+
+                SSLSocketFactory factory = context.getSocketFactory();
+                socket = factory.createSocket(ip,portNo);
+            }
             
             ipField.setEditable(false);
             portField.setEditable(false);
-            
-            connectButton.setText("Stop Listening");
-            connectButton.setMnemonic('S');
-            connectButton.setToolTipText("Stop Listening");
-            //sendButton.setEnabled(true);
-            //sendField.setEditable(true);
+            connectButton.setText("Disconnect");
+            connectButton.setMnemonic('D');
+            connectButton.setToolTipText("Stop Connection");
+            sendButton.setEnabled(true);
+            sendField.setEditable(true);
         } catch (Exception e) {
-            error(e.getMessage(), "Starting Server at "+portNo);
+			e.printStackTrace();
+            error(e.getMessage(), "Opening connection");
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
             return;
         }
-        //changeBorder(" "+socket.getInetAddress().getHostName()+
-        //	" ["+socket.getInetAddress().getHostAddress()+"] ");
+        changeBorder(" "+socket.getInetAddress().getHostName()+
+                " ["+socket.getInetAddress().getHostAddress()+"] ");
         setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-        messagesField.setText("> Server Started on Port: "+portNo+NEW_LINE);
-        append("> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
-        socketServer=SocketServer.handle(this,server);
-        //sendField.requestFocus();
+        messagesField.setText("");
+        socketClient=SocketClient.handle(this,socket);
+        sendField.requestFocus();
     }
-    //disconnect a client
+    
     public synchronized void disconnect() {
         try {
-            socketServer.setDesonnected(true);
-        } catch (Exception e) {}
-    }
-    
-    public synchronized void stop() {
-        try {
-            disconnect(); //close any client
-            socketServer.setStop(true);
-        } catch (Exception e) {}
-        server=null;
+            socketClient.setDesonnected(true);
+            socket.close();
+        } catch (Exception e) {
+            System.err.println("Error closing client : "+e);
+        }
+        socket=null;
+        out=null;
+        changeBorder(null);
         ipField.setEditable(true);
         portField.setEditable(true);
-        connectButton.setText("Start Listening");
-        connectButton.setMnemonic('S');
-        connectButton.setToolTipText("Start Listening");
-        append("> Server stopped");
-        append("> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
+        connectButton.setText("Connect");
+        connectButton.setMnemonic('C');
+        connectButton.setToolTipText("Start Connection");
+        sendButton.setEnabled(false);
+        sendField.setEditable(false);
     }
-    
-    public synchronized void setClientSocket(Socket s) {
-        
-        if(s==null) {
-            out=null;
-            socket = null;
-            changeBorder(null);
-            sendButton.setEnabled(false);
-            sendField.setEditable(false);
-            disconnectButton.setEnabled(false);
-        } else {
-            socket = s;
-            changeBorder(" "+socket.getInetAddress().getHostName()+
-                    " ["+socket.getInetAddress().getHostAddress()+"] ");
-            sendButton.setEnabled(true);
-            sendField.setEditable(true);
-            disconnectButton.setEnabled(true);
-        }
-    }
-    
     public void error(String error) {
         if(error==null || error.equals(""))
             return;
-        JOptionPane.showMessageDialog(SocketTestServer.this,
+        JOptionPane.showMessageDialog(SocketTestClient.this,
                 error, "Error", JOptionPane.ERROR_MESSAGE);
     }
     
     public void error(String error, String heading) {
         if(error==null || error.equals(""))
             return;
-        JOptionPane.showMessageDialog(SocketTestServer.this,
+        JOptionPane.showMessageDialog(SocketTestClient.this,
                 error, heading, JOptionPane.ERROR_MESSAGE);
     }
     
@@ -446,7 +441,7 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
         } catch (Exception e) {
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-            JOptionPane.showMessageDialog(SocketTestServer.this,
+            JOptionPane.showMessageDialog(SocketTestClient.this,
                     e.getMessage(),"Error Sending Message",
                     JOptionPane.ERROR_MESSAGE);
             disconnect();
@@ -456,15 +451,16 @@ public class SocketTestServer extends JPanel /*JFrame*/ {
     private void changeBorder(String ip) {
         if(ip==null || ip.equals(""))
             connectedBorder = BorderFactory.createTitledBorder(
-                    new EtchedBorder(), "Connected Client : < NONE >");
+                    new EtchedBorder(), "Connected To < NONE >");
         else
             connectedBorder = BorderFactory.createTitledBorder(
-                    new EtchedBorder(), "Connected Client : < "+ip+" >");
+                    new EtchedBorder(), "Connected To < "+ip+" >");
         CompoundBorder cb=new CompoundBorder(
                 BorderFactory.createEmptyBorder(5,10,10,10),
                 connectedBorder);
         centerPanel.setBorder(cb);
         invalidate();
         repaint();
-    }    
+    }
+    
 }
