+++ src/main/java/net/sf/sockettest/swing/SocketTestServer.java
@@ -3,24 +3,19 @@ package net.sf.sockettest.swing;
 import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
-import javax.swing.event.*;
 import javax.swing.border.*;
-
-import java.security.SecureRandom;
 
 import java.net.*;
 import java.io.*;
 
-import javax.net.*;
-import javax.net.ssl.*;
-
 import net.sf.sockettest.*;
+import org.apache.commons.lang3.StringEscapeUtils;
+
 /**
  *
  * @author Akshathkumar Shetty
  */
-public class SocketTestClient extends JPanel {
-    
+public class SocketTestServer extends JPanel /*JFrame*/ {
     private final String NEW_LINE = "\r\n";
     private ClassLoader cl = getClass().getClassLoader();
     public ImageIcon logo = new ImageIcon(
@@ -38,32 +33,33 @@ public class SocketTestClient extends JPanel {
     private JLabel portLabel = new JLabel("Port");
     private JLabel logoLabel = new JLabel("SocketTest v 3.0", logo,
             JLabel.CENTER);
-    private JTextField ipField = new JTextField("127.0.0.1",20);
+    private JTextField ipField = new JTextField("0.0.0.0",20);
     private JTextField portField = new JTextField("21",10);
     private JButton portButton = new JButton("Port");
-    private JButton connectButton = new JButton("Connect");
+    private JButton connectButton = new JButton("Start Listening");
     
-    private JLabel convLabel = new JLabel("Conversation with host");
-    private Border connectedBorder = BorderFactory.createTitledBorder(new EtchedBorder(),"Connected To < NONE >");
+    private JLabel convLabel = new JLabel("Conversation with Client");
+    private Border connectedBorder = BorderFactory.createTitledBorder(new EtchedBorder(),"Connected Client : < NONE >");
     private JTextArea messagesField = new JTextArea();
     
     private JLabel sendLabel = new JLabel("Message");
     private JTextField sendField = new JTextField();
     
     private JButton sendButton = new JButton("Send");
+    private JButton disconnectButton = new JButton("Disconnect");
     private JButton saveButton = new JButton("Save");
     private JButton clearButton = new JButton("Clear");
     
-    private JCheckBox secureButton = new JCheckBox("Secure");
-    private boolean isSecure = false;
     private GridBagConstraints gbc = new GridBagConstraints();
     
     private Socket socket;
+    private ServerSocket server;
+    private SocketServer socketServer;
     private PrintWriter out;
-    private SocketClient socketClient;
+    
     protected final JFrame parent;
     
-    public SocketTestClient(final JFrame parent) {
+    public SocketTestServer(final JFrame parent) {
         //Container cp = getContentPane();
         this.parent = parent;
         Container cp = this;
@@ -84,7 +80,7 @@ public class SocketTestClient extends JPanel {
         
         gbc.weightx = 1.0; //streach
         gbc.gridx = 1;
-        gbc.gridwidth = 4;
+        gbc.gridwidth = 3;
         gbc.fill = GridBagConstraints.HORIZONTAL;
         ActionListener ipListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -135,28 +131,12 @@ public class SocketTestClient extends JPanel {
         gbc.gridx = 3;
         gbc.gridwidth = 1;
         gbc.fill = GridBagConstraints.NONE;
-        connectButton.setMnemonic('C');
-        connectButton.setToolTipText("Start Connection");
+        connectButton.setMnemonic('S');
+        connectButton.setToolTipText("Start Listening");
         connectButton.addActionListener(connectListener);
         toPanel.add(connectButton, gbc);
         
-        
-        gbc.weightx = 0.0;
-        gbc.gridy = 1;
-        gbc.gridx = 4;
-        gbc.gridwidth = 1;
-        gbc.fill = GridBagConstraints.NONE;
-        secureButton.setToolTipText("Set Has Secure");
-        secureButton.addItemListener(
-                new ItemListener() {
-            public void itemStateChanged(ItemEvent e) {
-                isSecure = !isSecure;
-            }
-        });
-        toPanel.add(secureButton, gbc);
-        
-        
-        toPanel.setBorder(BorderFactory.createTitledBorder(new EtchedBorder(),"Connect To"));
+        toPanel.setBorder(BorderFactory.createTitledBorder(new EtchedBorder(),"Listen On"));
         topPanel.setLayout(new BorderLayout(10,0));
         topPanel.add(toPanel);
         logoLabel.setVerticalTextPosition(JLabel.BOTTOM);
@@ -193,7 +173,7 @@ public class SocketTestClient extends JPanel {
         gbc.weightx = 0.0;
         gbc.fill = GridBagConstraints.NONE;
         sendButton.setEnabled(false);
-        sendButton.setToolTipText("Send text to host");
+        sendButton.setToolTipText("Send text to client");
         ActionListener sendListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
                 String msg = sendField.getText();
@@ -201,8 +181,8 @@ public class SocketTestClient extends JPanel {
                     sendMessage(msg);
                 else {
                     int value = JOptionPane.showConfirmDialog(
-                            SocketTestClient.this,  "Send Blank Line ?",
-                            "Send Data To Server",
+                            SocketTestServer.this,  "Send Blank Line ?",
+                            "Send Data To Client",
                             JOptionPane.YES_NO_OPTION);
                     if (value == JOptionPane.YES_OPTION)
                         sendMessage(msg);
@@ -212,6 +192,16 @@ public class SocketTestClient extends JPanel {
         sendButton.addActionListener(sendListener);
         sendField.addActionListener(sendListener);
         sendPanel.add(sendButton, gbc);
+        ActionListener disconnectListener = new ActionListener() {
+            public void actionPerformed(ActionEvent e) {
+                disconnect();
+            }
+        };
+        gbc.gridx = 3;
+        disconnectButton.addActionListener(disconnectListener);
+        disconnectButton.setEnabled(false);
+        sendPanel.add(disconnectButton, gbc);
+        
         sendPanel.setBorder(
                 new CompoundBorder(
                 BorderFactory.createEmptyBorder(0,0,0,3),
@@ -233,7 +223,7 @@ public class SocketTestClient extends JPanel {
         gbc.gridx = 1;
         gbc.gridheight = 1;
         gbc.fill = GridBagConstraints.HORIZONTAL;
-        saveButton.setToolTipText("Save conversation with host to a file");
+        saveButton.setToolTipText("Save conversation with client to a file");
         saveButton.setMnemonic('S');
         ActionListener saveListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -245,13 +235,13 @@ public class SocketTestClient extends JPanel {
                 String fileName="";
                 JFileChooser chooser = new JFileChooser();
                 chooser.setCurrentDirectory(new File("."));
-                int returnVal = chooser.showSaveDialog(SocketTestClient.this);
+                int returnVal = chooser.showSaveDialog(SocketTestServer.this);
                 if(returnVal == JFileChooser.APPROVE_OPTION) {
                     fileName=chooser.getSelectedFile().getAbsolutePath();
                     try {
                         Util.writeFile(fileName,text);
                     } catch (Exception ioe) {
-                        JOptionPane.showMessageDialog(SocketTestClient.this,
+                        JOptionPane.showMessageDialog(SocketTestServer.this,
                                 ""+ioe.getMessage(),
                                 "Error saving to file..",
                                 JOptionPane.ERROR_MESSAGE);
@@ -262,7 +252,7 @@ public class SocketTestClient extends JPanel {
         saveButton.addActionListener(saveListener);
         buttonPanel.add(saveButton, gbc);
         gbc.gridy = 1;
-        clearButton.setToolTipText("Clear conversation with host");
+        clearButton.setToolTipText("Clear conversation with client");
         clearButton.setMnemonic('C');
         ActionListener clearListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
@@ -290,29 +280,28 @@ public class SocketTestClient extends JPanel {
     
         /*
         public static void main(String args[]) {
-                SocketTestClient client=new SocketTestClient();
-                client.setTitle("SocketTest Client");
-                //client.pack();
-                client.setSize(500,400);
-                Util.centerWindow(client);
-                client.setDefaultCloseOperation(EXIT_ON_CLOSE);
-                client.setIconImage(client.logo.getImage());
-                client.setVisible(true);
+                SocketTestServer stServer=new SocketTestServer();
+                stServer.setTitle("SocketTest Server");
+                stServer.setSize(500,400);
+                Util.centerWindow(stServer);
+                stServer.setDefaultCloseOperation(EXIT_ON_CLOSE);
+                stServer.setIconImage(stServer.logo.getImage());
+                stServer.setVisible(true);
         }
          */
     
-    /////////////////
-    //action methods
-    //////////////////
+    /////////////////////
+    //action & helper methods
+    /////////////////////
     private void connect() {
-        if(socket!=null) {
-            disconnect();
+        if(server!=null) {
+            stop();
             return;
         }
         String ip=ipField.getText();
         String port=portField.getText();
         if(ip==null || ip.equals("")) {
-            JOptionPane.showMessageDialog(SocketTestClient.this,
+            JOptionPane.showMessageDialog(SocketTestServer.this,
                     "No IP Address. Please enter IP Address",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             ipField.requestFocus();
@@ -320,7 +309,7 @@ public class SocketTestClient extends JPanel {
             return;
         }
         if(port==null || port.equals("")) {
-            JOptionPane.showMessageDialog(SocketTestClient.this,
+            JOptionPane.showMessageDialog(SocketTestServer.this,
                     "No Port number. Please enter Port number",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             portField.requestFocus();
@@ -329,7 +318,7 @@ public class SocketTestClient extends JPanel {
         }
         setCursor(Cursor.getPredefinedCursor(Cursor.WAIT_CURSOR));
         if(!Util.checkHost(ip)) {
-            JOptionPane.showMessageDialog(SocketTestClient.this,
+            JOptionPane.showMessageDialog(SocketTestServer.this,
                     "Bad IP Address",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             ipField.requestFocus();
@@ -341,7 +330,7 @@ public class SocketTestClient extends JPanel {
         try	{
             portNo=Integer.parseInt(port);
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(SocketTestClient.this,
+            JOptionPane.showMessageDialog(SocketTestServer.this,
                     "Bad Port number. Please enter Port number",
                     "Error connecting", JOptionPane.ERROR_MESSAGE);
             portField.requestFocus();
@@ -350,72 +339,86 @@ public class SocketTestClient extends JPanel {
             return;
         }
         try {
-            if(isSecure==false) {
-                System.out.println("Connectig in normal mode : "+ip+":"+portNo);
-                socket = new Socket(ip,portNo);
-            } else {
-                System.out.println("Connectig in secure mode : "+ip+":"+portNo);
-                //SocketFactory factory = SSLSocketFactory.getDefault();
-				
-				TrustManager[] tm = new TrustManager[] { new MyTrustManager(SocketTestClient.this) }; 
-
-                SSLContext context = SSLContext.getInstance("TLS");
-                context.init(new KeyManager[0], tm, new SecureRandom());
-
-                SSLSocketFactory factory = context.getSocketFactory();
-                socket = factory.createSocket(ip,portNo);
-            }
+            InetAddress bindAddr=null;
+            if(!ip.equals("0.0.0.0"))
+                bindAddr = InetAddress.getByName(ip);
+            else
+                bindAddr = null;
+            server = new ServerSocket(portNo,1,bindAddr);
             
             ipField.setEditable(false);
             portField.setEditable(false);
-            connectButton.setText("Disconnect");
-            connectButton.setMnemonic('D');
-            connectButton.setToolTipText("Stop Connection");
-            sendButton.setEnabled(true);
-            sendField.setEditable(true);
+            
+            connectButton.setText("Stop Listening");
+            connectButton.setMnemonic('S');
+            connectButton.setToolTipText("Stop Listening");
+            //sendButton.setEnabled(true);
+            //sendField.setEditable(true);
         } catch (Exception e) {
-			e.printStackTrace();
-            error(e.getMessage(), "Opening connection");
+            error(e.getMessage(), "Starting Server at "+portNo);
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
             return;
         }
-        changeBorder(" "+socket.getInetAddress().getHostName()+
-                " ["+socket.getInetAddress().getHostAddress()+"] ");
+        //changeBorder(" "+socket.getInetAddress().getHostName()+
+        //	" ["+socket.getInetAddress().getHostAddress()+"] ");
         setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-        messagesField.setText("");
-        socketClient=SocketClient.handle(this,socket);
-        sendField.requestFocus();
+        messagesField.setText("> Server Started on Port: "+portNo+NEW_LINE);
+        append("> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
+        socketServer=SocketServer.handle(this,server);
+        //sendField.requestFocus();
     }
-    
+    //disconnect a client
     public synchronized void disconnect() {
         try {
-            socketClient.setDesonnected(true);
-            socket.close();
-        } catch (Exception e) {
-            System.err.println("Error closing client : "+e);
-        }
-        socket=null;
-        out=null;
-        changeBorder(null);
+            socketServer.setDesconnected(true);
+        } catch (Exception e) {}
+    }
+    
+    public synchronized void stop() {
+        try {
+            disconnect(); //close any client
+            socketServer.setStop(true);
+        } catch (Exception e) {}
+        server=null;
         ipField.setEditable(true);
         portField.setEditable(true);
-        connectButton.setText("Connect");
-        connectButton.setMnemonic('C');
-        connectButton.setToolTipText("Start Connection");
-        sendButton.setEnabled(false);
-        sendField.setEditable(false);
+        connectButton.setText("Start Listening");
+        connectButton.setMnemonic('S');
+        connectButton.setToolTipText("Start Listening");
+        append("> Server stopped");
+        append("> ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
     }
+    
+    public synchronized void setClientSocket(Socket s) {
+        
+        if(s==null) {
+            out=null;
+            socket = null;
+            changeBorder(null);
+            sendButton.setEnabled(false);
+            sendField.setEditable(false);
+            disconnectButton.setEnabled(false);
+        } else {
+            socket = s;
+            changeBorder(" "+socket.getInetAddress().getHostName()+
+                    " ["+socket.getInetAddress().getHostAddress()+"] ");
+            sendButton.setEnabled(true);
+            sendField.setEditable(true);
+            disconnectButton.setEnabled(true);
+        }
+    }
+    
     public void error(String error) {
         if(error==null || error.equals(""))
             return;
-        JOptionPane.showMessageDialog(SocketTestClient.this,
+        JOptionPane.showMessageDialog(SocketTestServer.this,
                 error, "Error", JOptionPane.ERROR_MESSAGE);
     }
     
     public void error(String error, String heading) {
         if(error==null || error.equals(""))
             return;
-        JOptionPane.showMessageDialog(SocketTestClient.this,
+        JOptionPane.showMessageDialog(SocketTestServer.this,
                 error, heading, JOptionPane.ERROR_MESSAGE);
     }
     
@@ -443,7 +446,7 @@ public class SocketTestClient extends JPanel {
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
         } catch (Exception e) {
             setCursor(Cursor.getPredefinedCursor(Cursor.DEFAULT_CURSOR));
-            JOptionPane.showMessageDialog(SocketTestClient.this,
+            JOptionPane.showMessageDialog(SocketTestServer.this,
                     e.getMessage(),"Error Sending Message",
                     JOptionPane.ERROR_MESSAGE);
             disconnect();
@@ -453,16 +456,15 @@ public class SocketTestClient extends JPanel {
     private void changeBorder(String ip) {
         if(ip==null || ip.equals(""))
             connectedBorder = BorderFactory.createTitledBorder(
-                    new EtchedBorder(), "Connected To < NONE >");
+                    new EtchedBorder(), "Connected Client : < NONE >");
         else
             connectedBorder = BorderFactory.createTitledBorder(
-                    new EtchedBorder(), "Connected To < "+ip+" >");
+                    new EtchedBorder(), "Connected Client : < "+ip+" >");
         CompoundBorder cb=new CompoundBorder(
                 BorderFactory.createEmptyBorder(5,10,10,10),
                 connectedBorder);
         centerPanel.setBorder(cb);
         invalidate();
         repaint();
-    }
-    
+    }    
 }
