+++ src/main/java/net/sf/sockettest/swing/SocketTestUdp.java
@@ -3,7 +3,6 @@ package net.sf.sockettest.swing;
 import java.awt.*;
 import java.awt.event.*;
 import javax.swing.*;
-import javax.swing.event.*;
 import javax.swing.border.*;
 
 import java.net.*;
