+++ src/main/resources/readme.txt
